readdir: common Readdir code
